pyproject.toml
[project]
name = "projeto-1-soa"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas Santos Silva",email = "luss1010@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio-tools (>=1.76.0,<2.0.0)",
    "fastapi[standard] (>=0.121.1,<0.122.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
